backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
pymongo==4.5.0
pyjwt==2.8.0
bcrypt==4.0.1
qrcode==7.4.2
pillow==11.0.0
requests==2.31.0
razorpay==1.3.0
cryptography==41.0.3
